backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
numpy==1.26.2
pandas==2.1.4
